AnimatedBellota incorporated into Bellota using the same IndirectTexture class. Texture is now a variant representing IndirectTexture or DirectTexture. (#9)
Now we can use both color schemes to specify colors in bellota's textures, direct and indirect.

- Previous Texture definition is now IndirectTexture, which as keeps a color pallete and a matrix of indices to those colors.
- New definition for DirectTexture, where each element of the matrix is a specific collor.
- Texture is now defined as std::variant<IndirectTexture, DirectTexture> so it can be used a both types.
- Some Visitors are implemented to handle basic and common Texture functionality.
- AnimatedBellota was eliminated and its functionality is now part of standard Bellota.
- Animations are only supported with Indirect Textures via its layers.
- Text rendering is only supported with Indirect Textures.
- Examples have been renamed to better illustrate their intent.

Bellota's texture is now implemented via an OpenGL Texture Array 2D instead of a simple Texture 2D. This is mostly to decrease engine complexity and support a single flow. Much of the AnimatedBellota was duplicated. Bellota has a reference to a Texture, so it can use Direct or Indirect color schemes.

https://github.com/dantros/nothofagus/issues/6